pyproject.toml
[tool.poetry]
name = "kwenta-bot"
version = "0.1.0"
description = ""
authors = ["Kenneth V. Domingo <hello@kvdomingo.xyz>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.1.0"
loguru = "^0.6.0"
google-api-python-client = "^2.73.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.8.0"
tabulate = "^0.9.0"

[tool.poetry.group.dev.dependencies]
google-cloud-secret-manager = "^2.15.0"
mako = "^1.2.4"
python-dotenv = "^0.21.1"
watchdog = {extras = ["watchmedo"], version = "^2.2.1"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.isort]
profile = "black"
